Playlist 조회 api 수정 및 조회 시 userNo request header 추가

+++ src/main/java/com/example/demo/model/SearchSong.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class SearchSong {
     private String title;
     private Long id;
-    private short track;
-    private short length;
+    private Integer track;
+    private Integer length;
 
 }
